// src/rpg-multiplayer-game/src/client/src/consts/AvatarKeys.ts
enum AvatarKeys {
  ADRIAN_TOMKINS = 'adrian-tomkins',
  CALVIN_KOEPKE = 'calvin-koepke',
  JOEL_BOWEN = 'joel-bowen',
  JOHN_GEIGER = 'john-geiger',
  LAKSHMI_VEGESNA = 'lakshmi-vegesna',
  MICHAEL_KANTOR = 'michael-kantor',
  NICK_TAYLOR = 'nick-taylor',
  NICOLAS_DEBIN = 'nicolas-debin',
  RYAN_MCKAY = 'ryan-mckay',
};

export default AvatarKeys